refactoring a little bit
