+++ conventions/tests/services/test_selection_service.py
@@ -14,7 +14,7 @@ from conventions.services import (
 )
 from core.tests import utils_fixtures
 from instructeurs.models import Administration
-from programmes.subforms.lot_selection import (
+from conventions.forms.lot_selection import (
     ProgrammeSelectionFromDBForm,
     ProgrammeSelectionFromZeroForm,
 )
